Fix generator
